change folder name to Docs
add docs:build to package.json

Small updates

Update gitignore

Update

update eslintignore

update author
