+++ .eslintignore
@@ -1,3 +1,3 @@
 **/build/*
 calc
-/openapi
+docs
