+++ docs/index.html
@@ -1,0 +1,14 @@
+<!doctype html>
+<html>
+
+<head>
+  <meta charset="UTF-8">
+  <title>Substrate API Sidecar OpenApi</title>
+  <link rel="icon" href="favicon.ico">
+</head>
+
+<body>
+  <main id="spec"></main>
+</body>
+
+</html>
