--- openapi/README.md
@@ -1,3 +1,0 @@
-# Sidecar OpenApi UI
-
-Proposed OpenApi spec for [substrate-api-sidecar](https://github.com/paritytech/substrate-api-sidecar) v1.
