--- openapi/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html>
-
-<head>
-  <meta charset="UTF-8">
-  <title>Substrate API Sidecar OpenApi</title>
-  <link rel="icon" href="favicon.ico">
-</head>
-
-<body>
-  <main id="spec"></main>
-</body>
-
-</html>
