+++ package.json
@@ -9,6 +9,7 @@
     "build": "tsc",
     "build:calc": "bash ./calc/build.sh",
     "build:docker": "docker build -t substrate-api-sidecar .",
+    "build:docs": "(cd docs && yarn && yarn build)",
     "main": "node ./build/src/main.js",
     "lint": "tsc && eslint . --ext ts",
     "start": "yarn run build && yarn run main",
