+++ tsconfig.json
@@ -27,7 +27,7 @@
   ],
   "exclude": [
     "node_modules",
-    "openapi"
+    "docs"
   ],
   "include": [
     "./src/**/*.ts",
